Refactoring, extract remove locals function to new file
